+++ src/SoundPlayers/SoundPlayerInterface.h
@@ -7,5 +7,5 @@ class SoundPlayerInterface
 {
     public:
         void virtual init() = 0;
-        void virtual playFile(uint8_t fileNumber) = 0;
+        void virtual playFile(uint16_t fileNumber) = 0;
 };
